+++ tests/test_npz.py
@@ -1,8 +1,10 @@
 import numpy as np
 from time import sleep
+import os
 from shutil import rmtree
 from cache_decorator import Cache
 from .utils import standard_test_arrays
+
 
 @Cache(
     cache_path="{cache_dir}/{_hash}.npz",
@@ -13,6 +15,7 @@ def cached_function_single(a):
     sleep(2)
     return np.array([1, 2, 3])
 
+
 @Cache(
     cache_path="{cache_dir}/{_hash}.npz",
     cache_dir="./test_cache",
@@ -21,6 +24,7 @@ def cached_function_single(a):
 def cached_function_tuple(a):
     sleep(2)
     return np.array([1, 2, 3]), np.array([1, 2, 4])
+
 
 @Cache(
     cache_path="{cache_dir}/{_hash}.npz",
@@ -31,6 +35,7 @@ def cached_function_list(a):
     sleep(2)
     return [np.array([1, 2, 3]), np.array([1, 2, 4])]
 
+
 @Cache(
     cache_path="{cache_dir}/{_hash}.npz",
     cache_dir="./test_cache",
@@ -38,20 +43,36 @@ def cached_function_list(a):
 )
 def cached_function_dict(a):
     sleep(2)
-    return {"a":np.array([1, 2, 3]), "b":np.array([1, 2, 4])}
+    return {"a": np.array([1, 2, 3]), "b": np.array([1, 2, 4])}
+
 
 def test_npz_single():
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
     standard_test_arrays(cached_function_single)
-    rmtree("./test_cache")
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
+
 
 def test_npz_tuple():
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
     standard_test_arrays(cached_function_tuple)
-    rmtree("./test_cache")
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
+
 
 def test_npz_list():
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
     standard_test_arrays(cached_function_list)
-    rmtree("./test_cache")
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
+
 
 def test_npz_dict():
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
     standard_test_arrays(cached_function_dict)
-    rmtree("./test_cache")
+    if os.path.exists("./test_cache"):
+        rmtree("./test_cache")
